requirements.txt
# Python dependencies 
ursina==5.2.0
numpy==1.26.3 
pillow==10.2.0 